Updated Uniqueness Assurance to be sensitive to analysis time out, and to issue warnings if analysis takes a long time (but doesn't time out).
Also fixed the chain of evidence of @unique parameters.  IT used to be that these depended on the control flow of the method that contains the parameeter.  This is not correct.  They depend on the control-flows of the methods that invoke the method with the annotated parameter.

Removed some dead code.
 

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@17852 d02ae147-ee12-4a35-9452-91c89e8b2592