Fixed the method control flow promise drop to simply be a field of the StoreLattice: we know there is only one method, the one currently being assured.
Fixed XNotY to have proper equals() and hashCode() emthods.  This eliminates dupclies in the results that come about because of the way that instance initializers are handled in the control flow graph.  I cannot figure out a way at this point to prevent the generation of duplicate results, only the detection of them.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@18061 d02ae147-ee12-4a35-9452-91c89e8b2592