ASSIGNED - bug 1716: Update effects analysis https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1716
Updated effects analysis to ignore effects on objects referenced by immutable references.

They now create effects on a new "empty" target.  This allows me to report a result that can seen by the user.

Fixed Effects.getDeclaredMethodEffects() to return a list instead of a set.  This is necessary for deterministic assurance results.  Otherwise, we cannot be sure of the order in which actual effects are compared against the declared effects.  Became an issue once I added the empty target because now it is easy to have an implementation effect that satisfied many declared effects.

Should have return a list here all along.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@18790 d02ae147-ee12-4a35-9452-91c89e8b2592