REOPENED - bug 75: Reporting of reads and writes in effects results, even if masked or final http://surelogic.com/bugzilla/show_bug.cgi?id=75
Finished making the targets of effects from Anonymous Class Expressions have
special evidence that refers to effect in the ACE initialization from which
they are derived.  Obviously I had to add a new evidence class.  I had to add
some ugliness to the Effect and Target classes to make this work though:
Effect.changeSource() and Target.changeEvidence().  Effect.setSource() will
soon be removed when I fix the evidence for effects that come from method
calls.

For some reason this fixed missing effects (that I was not aware were missing) in the TypeAnnotationsAndEnums tests.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@18886 d02ae147-ee12-4a35-9452-91c89e8b2592