REOPENED - bug 75: Reporting of reads and writes in effects results, even if masked or final http://surelogic.com/bugzilla/show_bug.cgi?id=75
Add new CallEvidence (and subclasses) to describe targets derived from declared
and instantiated method effects.

Now only the actual formal and actual parameters involved for the particular
effect are listed.  Before, when we added the parameter mapping after the fact
in EffectAnalysis, we listed the entire mapping, which was confusing.

Added special evidence for when effects on qualified receivers are converted to
any instance effects.  Similarly, added special evidence for the Anonymous
Class Expression case where effects are unknown objects are converted to any
instance effects.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@18890 d02ae147-ee12-4a35-9452-91c89e8b2592